Moved methods from IdeaController into more specific classes; ContextMenuController and ContextMenuActions.
